docs: Clean up and align Showcase thumbnail images
Some of the thumbnails in Showcase were out of place
because of several irregularly sized thumbnails,
and some of them almost 300KB in filesize.

Resize them all to 600x400 (pixels), and use the `-tn.png`
suffix.

When necessary, the website snapshot is re-captured using
gnome-web-photo.  Then, the following commands (or a combination
thereof) are used to crop and resize the image into a thumbnail,
and to reduce its filesize:

    $ convert example.png -crop 900x600+0+0 \
                          -filter Lanczos2Sharp -distort Resize 600x400 \
                          example-tn.png
    $ pngquant --nofs -v --speed 1 --quality 65-80 example-tn.png
    $ optipng -o7 -zm1-9 example-tn-or8.png
    $ mv example-tn-or8.png example-tn.png